requirements.txt

streamlit
transformers
torch
google-generativeai